src/utils.py
from datetime import datetime

from google.protobuf.timestamp_pb2 import Timestamp
import pytz

utc = pytz.UTC


def timestamp_from_datetime(dt: datetime):
    pb_ts = Timestamp()
    pb_ts.FromDatetime(dt)
    return pb_ts

def to_aware_datetime(ts: Timestamp):
    """
    Turns a protobuf Timestamp object into a timezone-aware datetime
    """
    return utc.localize(ts.ToDatetime())
